Implementar registro completo de usuario y empresa con relación empresa_usuario
- Crear RegistroService para manejar registro de usuarios y empresas
- Crear EmpresaUsuarioRepository para gestionar relaciones empresa-usuario
- Agregar mutation register en AuthMutation con generación de JWT
- Actualizar tipos GraphQL (RegisterInput, RegisterResponse, UserRegistered)
- Corregir tipos de datos en EmpresaUsuario y EmpresaUsuarioId (Long)
- Corregir query JPQL en EmpresaRepository
- Ajustar métodos en EmpresaUsuarioRepository para usar campos primitivos
- Crear auditorías para usuario, empresa y relación
- Registrar sesión inicial después del registro
- Revertir ddl-auto a validate en application.yml

+++ src/main/java/com/vias/uc/backend/model/EmpresaUsuarioId.java
@@ -10,12 +10,12 @@ import java.util.Objects;
 public class EmpresaUsuarioId implements Serializable {
 
     private Integer empresa;  // Debe coincidir con el nombre del atributo en EmpresaUsuario
-    private Integer usuario;  // Debe coincidir con el nombre del atributo en EmpresaUsuario (INTEGER en BD)
+    private Long usuario;  // Debe coincidir con BIGINT en BD y Long en EmpresaUsuario
 
     public EmpresaUsuarioId() {
     }
 
-    public EmpresaUsuarioId(Integer empresa, Integer usuario) {
+    public EmpresaUsuarioId(Integer empresa, Long usuario) {
         this.empresa = empresa;
         this.usuario = usuario;
     }
@@ -28,11 +28,11 @@ public class EmpresaUsuarioId implements Serializable {
         this.empresa = empresa;
     }
 
-    public Integer getUsuario() {
+    public Long getUsuario() {
         return usuario;
     }
 
-    public void setUsuario(Integer usuario) {
+    public void setUsuario(Long usuario) {
         this.usuario = usuario;
     }
